+++ .github/workflows/update_llm_perf_cuda_pytorch.yaml
@@ -53,4 +53,4 @@ jobs:
             pip install packaging && pip install flash-attn einops scipy auto-gptq optimum bitsandbytes autoawq codecarbon
             pip install -U transformers huggingface_hub[hf_transfer]
             pip install -e .
-            python llm_perf/update_llm_perf_cuda_pytorch.py
+            python llm_perf/benchmarks/update_llm_perf_cuda_pytorch.py
